Move `ScanEvent`, `FatalShutdown`, `resolve_seeds` to `lib`
